Attempt to fix but with magic.t for which I have no details
Got a report that Magic.t fails because of missing
Devel::Declare::Parser. No further details. I have guessed at the cause
and this would fix it if that is actually the cause.

I am assuming that the problem is with systems that already had
Exporter::Declare 0.106, but never used -magic, as such they did have an
Exporter::Declare::Magic module for the test to load, and so it did. But
they also did not have Devel::Declare::Parser, so loading Magic.pm
failed.

This will now check for Devel::Declare::Parser as well as
Exporter::Declare::Magic (in order). The test will be skipped if either
are missing.